+++ src/engine/rendering/Sprite.java
@@ -42,7 +42,7 @@ public class Sprite extends Entity
     private float timeSinceLastFrame;
     private Vector2f[] spriteSheetFrameOffsets;
 
-    public Sprite(String name, short HIERARCHY_INDEX, String textureAssetKey, Vector4f mainTextureTint, Vector2f mainTextureOffset, Vector2f mainTextureScale)
+    public Sprite(String name, int HIERARCHY_INDEX, String textureAssetKey, Vector4f mainTextureTint, Vector2f mainTextureOffset, Vector2f mainTextureScale)
     {
         super(name, EntityType.Sprite, HIERARCHY_INDEX);
         this.shader = Scene.standard2dShader;
